Changed implementation to one that does not useSWR
